refactor: refactor package to use src folder
