+++ cli/parse-badge-json-config-file.ts
@@ -3,7 +3,7 @@ import type { Params } from '../types';
 import { validateConfig } from './config-validation';
 import path from 'path';
 import chalk from 'chalk';
-import { addBadge } from '../index';
+import { addBadge } from '../src/index';
 
 export function parseBadgeJsonConfigFile(filePath: string) {
   const resolvedPath = path.resolve(filePath);
