+++ src/assets/fonts/open-sans-128-black.png
