+++ src/assets/ribbon-overlay-adaptive.png
