+++ src/assets/ribbon-overlay.png
