+++ src/get-result-path.ts
